actualizacao de codigo- 1071223 1100669 1111112

+++ eapli.expensemanager/src/Persistance/JPA/ExpenseTypeJpa.java
@@ -13,7 +13,7 @@ import Persistance.IExpenseTypeRepository;
  *
  * @author Rui Damiao
  */
-public class ExpenseTypeJpa extends JpaGeneric<ExpenseType, Long> implements IExpenseTypeRepository{
+public class ExpenseTypeJpa extends JpaGeneric<ExpenseType, Integer> implements IExpenseTypeRepository{
       
     @Override
       public void saveExpenseType(ExpenseType eT){
